# [GAPRINDASHVILI] Faster report result index pages